fix: comprehensive security audit remediation (97 findings, 75 fixed)
14-round security audit addressing 97 findings across 148 files:

Security fixes:
- CRIT-1: Reject empty MCP bearer tokens (authentication bypass)
- HIGH-1..12: CSP nonce, HMAC replay, CSRF defaults, SSE/gRPC bounds,
  shadow proxy race, bot detection limits, circuit breaker leak, CI pinning
- Race conditions: conn/pool, cluster callbacks, backend manager, health passive
- Resource exhaustion: SSE/gRPC/io.Copy limits, coalesce/cache/retry bounds,
  MCP SSE clients, TCP/UDP connection limits, compression buffers
- Integer overflow: ParseByteSize, parseCustomDuration, TruncateDuration,
  rate limiter div-by-zero, config decoders, payload length framing,
  balancer index, metrics counter shard, weight bounds
- Error handling: cluster Apply errors, UDP drops, WebSocket write deadline,
  PROXY protocol ports, MCP error leakage, config decoder panics
- Type safety: unchecked atomic.Value assertions (proxy, backend, router),
  proxy.go panic guards with safe fallbacks
- Secrets zeroing: ZeroSecrets() on apikey/basic middleware, cluster auth,
  engine shutdown integration via SecretZeroer interface
- Goroutine lifecycle: WaitGroup for SNI/TCP listeners, config watcher,
  shadow proxy semaphore, cache revalidation timeout
- Supply chain: pinned CI tools (gosec, golangci-lint, staticcheck, benchstat,
  nancy), Dockerfile image digests, SBOM generation, govulncheck

WebUI audit:
- Eliminated all `any` types, added proper interfaces
- Added compliance, CI, and infrastructure fixes

Tests: 7 new ZeroSecrets tests, updated integration test auth headers

+++ internal/metrics/counter.go
@@ -61,7 +61,7 @@ func (c *Counter) Inc() {
 
 // Add adds n to the counter.
 func (c *Counter) Add(n int64) {
-	shard := int(gidCounter.Add(1)) & shardMask
+	shard := int(uint64(gidCounter.Add(1)) & uint64(shardMask))
 	c.shards[shard].Add(n)
 }
